% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepsouth.R
\docType{data}
\name{deepsouth}
\alias{deepsouth}
\title{Deep South Network}
\format{A data frame with 18 rows and 14 variables}
\source{
Davies et al (1941) Deep South: A sociological anthropological study of caste and class. University of Chicago Press.
}
\usage{
deepsouth
}
\description{
This is a data set of 18 women observed over a nine-month period. During that period, various subsets of these women met in a series of 14 informal social events. The data recored which women met for which events.
}
\keyword{datasets}
