% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.manet.R
\name{print.manet}
\alias{print.manet}
\title{Printing the output from the multiple allocation clustering}
\usage{
\method{print}{manet}(x, digits = 3, ...)
}
\arguments{
\item{x}{A manet object.}

\item{digits}{Number of digits. Default is 3.}

\item{...}{Additional arguments to the print function.}
}
\description{
This function prints the output of the manet function
}
\examples{
data(deepsouth)
ds<-manet(deepsouth,K=2,maxT=100)
print(ds)
}
