% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.manet.R
\name{summary.manet}
\alias{summary.manet}
\title{Summarising the output from the multiple allocation clustering}
\usage{
\method{summary}{manet}(object, digits = 3, ...)
}
\arguments{
\item{object}{A manet object.}

\item{digits}{Number of digits. Default is 3.}

\item{...}{Additional arguments to the summary function.}
}
\description{
This function summarises the output of the manet function
}
\examples{
data(deepsouth)
ds<-manet(deepsouth,K=2,maxT=100)
summary(ds)
}
